Create checkSet.rb
Working on implementing a method to check if user entered input for a choice of cards is a set.

As of now if works for checking the color only. It should be the same exact logic for the rest, but its only printing true so far for shape and number, so I'm working on fixing that. 

Just wanted to keep everyone updated so let me know if you have questions. To run, use the card class and makeSet function Roman wrote. 

The rules for a set are on wikipeida